requirements.txt
Django==3.1b1

django-ninja==0.2.0
pydantic==1.5.1

djangorestframework==3.11.0

Flask==1.1.2
marshmallow==3.6.1

sanic==20.6.3

uvicorn==0.11.5
uWSGI==2.0.19.1

aiohttp==3.6.2

requests
